+++ app/src/main/java/myapplication/second/workinghourmanagement/store/notice/OwnerNoticeDetailActivity.kt
@@ -1,4 +1,4 @@
-package myapplication.second.workinghourmanagement.store
+package myapplication.second.workinghourmanagement.store.notice
 
 import android.content.Context
 import android.content.Intent
