+++ app/src/main/java/myapplication/second/workinghourmanagement/store/notice/OwnerNoticeListAdapter.kt
@@ -1,4 +1,4 @@
-package myapplication.second.workinghourmanagement.store
+package myapplication.second.workinghourmanagement.store.notice
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -6,20 +6,20 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import myapplication.second.workinghourmanagement.databinding.ItemOwnerNoticeBinding
-import myapplication.second.workinghourmanagement.dto.ResultGetNotice
+import myapplication.second.workinghourmanagement.dto.notice.ResponseGetNotice
 
 class OwnerNoticeListAdapter(
-    private val onClick: (ResultGetNotice) -> Unit
-): ListAdapter<ResultGetNotice, OwnerNoticeListAdapter.OwnerNoticeListViewHolder>(DiffCallback()) {
+    private val onClick: (ResponseGetNotice) -> Unit
+): ListAdapter<ResponseGetNotice, OwnerNoticeListAdapter.OwnerNoticeListViewHolder>(DiffCallback()) {
 
-    private val noticeList = mutableListOf<ResultGetNotice>()
+    private val noticeList = mutableListOf<ResponseGetNotice>()
 
     class OwnerNoticeListViewHolder(
         val binding: ItemOwnerNoticeBinding,
-        private val onClick: (ResultGetNotice) -> Unit
+        private val onClick: (ResponseGetNotice) -> Unit
     ) : RecyclerView.ViewHolder(binding.root) {
 
-        fun bind(noticeItem: ResultGetNotice) {
+        fun bind(noticeItem: ResponseGetNotice) {
 
             binding.root.setOnClickListener {
                 onClick(noticeItem)
@@ -30,13 +30,13 @@ class OwnerNoticeListAdapter(
         }
     }
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): OwnerNoticeListAdapter.OwnerNoticeListViewHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): OwnerNoticeListViewHolder {
         val binding = ItemOwnerNoticeBinding.inflate(LayoutInflater.from(parent.context), parent, false)
 
-        return OwnerNoticeListAdapter.OwnerNoticeListViewHolder(binding, onClick)
+        return OwnerNoticeListViewHolder(binding, onClick)
     }
 
-    override fun onBindViewHolder(holder: OwnerNoticeListAdapter.OwnerNoticeListViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: OwnerNoticeListViewHolder, position: Int) {
         val noticeItem = noticeList[position]
 
         holder.bind(noticeItem)
@@ -46,29 +46,29 @@ class OwnerNoticeListAdapter(
         return noticeList.size
     }
 
-    fun setNoticeList(noticeList: List<ResultGetNotice>) {
+    fun setNoticeList(noticeList: List<ResponseGetNotice>) {
 //        noticeList.clear()
 //        noticeList.addAll(noticeList)
         notifyDataSetChanged()
     }
 
-    fun add(position: Int, notice: ResultGetNotice) {
+    fun add(position: Int, notice: ResponseGetNotice) {
         noticeList.add(position, notice)
         notifyItemInserted(position)
     }
 
-    fun replaceItem(notice: ResultGetNotice) {
+    fun replaceItem(notice: ResponseGetNotice) {
         val index = noticeList.indexOf(notice)
         noticeList[index] = notice
         notifyItemChanged(index)
     }
 
-    private class DiffCallback : DiffUtil.ItemCallback<ResultGetNotice>(){
-        override fun areItemsTheSame(oldItem: ResultGetNotice, newItem: ResultGetNotice): Boolean {
+    private class DiffCallback : DiffUtil.ItemCallback<ResponseGetNotice>(){
+        override fun areItemsTheSame(oldItem: ResponseGetNotice, newItem: ResponseGetNotice): Boolean {
             return oldItem.noticeId == newItem.noticeId
         }
 
-        override fun areContentsTheSame(oldItem: ResultGetNotice, newItem: ResultGetNotice): Boolean {
+        override fun areContentsTheSame(oldItem: ResponseGetNotice, newItem: ResponseGetNotice): Boolean {
             return oldItem == newItem
         }
     }
