+++ app/src/main/java/myapplication/second/workinghourmanagement/store/schedule/OwnerImportantScheduleListAdapter.kt
@@ -1,4 +1,4 @@
-package myapplication.second.workinghourmanagement.store
+package myapplication.second.workinghourmanagement.store.schedule
 
 import androidx.recyclerview.widget.RecyclerView
 import myapplication.second.workinghourmanagement.databinding.ItemOwnerImportantScheduleBinding
