+++ app/src/main/java/myapplication/second/workinghourmanagement/store/schedule/OwnerModifyScheduleActivity.kt
@@ -1,4 +1,4 @@
-package myapplication.second.workinghourmanagement.store
+package myapplication.second.workinghourmanagement.store.schedule
 
 import android.content.Context
 import android.content.Intent
@@ -7,17 +7,19 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import myapplication.second.workinghourmanagement.R
 import myapplication.second.workinghourmanagement.RetrofitService
-import myapplication.second.workinghourmanagement.databinding.ActivityOwnerPostScheduleBinding
+import myapplication.second.workinghourmanagement.databinding.ActivityOwnerModifyScheduleBinding
+import myapplication.second.workinghourmanagement.store.todo_list.BottomSheetSetDate
+import myapplication.second.workinghourmanagement.store.todo_list.BottomSheetSetTime
 
-class OwnerPostScheduleActivity: AppCompatActivity() {
+class OwnerModifyScheduleActivity: AppCompatActivity() {
 
-    private lateinit var binding: ActivityOwnerPostScheduleBinding
+    private lateinit var binding: ActivityOwnerModifyScheduleBinding
     private lateinit var service: RetrofitService
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        binding = DataBindingUtil.setContentView(this, R.layout.activity_owner_post_schedule)
+        binding = DataBindingUtil.setContentView(this, R.layout.activity_owner_modify_schedule)
         binding.lifecycleOwner = this
 
         setupView()
@@ -25,7 +27,7 @@ class OwnerPostScheduleActivity: AppCompatActivity() {
     }
 
     private fun setupView() {
-
+        TODO("Not yet implemented")
     }
 
     private fun setupListeners() {
@@ -45,24 +47,19 @@ class OwnerPostScheduleActivity: AppCompatActivity() {
             openSetTimeBottomSheetDialogFragment()
         }
 
-        binding.btnPostSchedule.setOnClickListener {
-            postSchedule()
+        binding.btnModifySchedule.setOnClickListener {
+            // TODO: 스케줄 변경할 때의 API 연동 구현
         }
     }
 
-    private fun postSchedule() {
-        // TODO: 일정 등록할 때 필요한 API 구현
-        val scheduleInfo = HashMap<String, String>()
-    }
-
     private fun openSetDateBottomSheetDialogFragment() {
-        val bottomSheet = BottomSheetSetDate()
-        bottomSheet.show(supportFragmentManager, bottomSheet.tag)
+        val intent = BottomSheetSetDate.getIntent(this)
+        startActivity(intent)
     }
 
     private fun openSetTimeBottomSheetDialogFragment() {
-        val bottomSheet = BottomSheetSetTime()
-        bottomSheet.show(supportFragmentManager, bottomSheet.tag)
+        val intent = BottomSheetSetTime.getIntent(this)
+        startActivity(intent)
     }
 
     private fun intentSelectColorPalette() {
@@ -72,6 +69,6 @@ class OwnerPostScheduleActivity: AppCompatActivity() {
 
     companion object {
         fun getIntent(context: Context) =
-            Intent(context, OwnerPostScheduleActivity::class.java)
+            Intent(context, OwnerModifyScheduleActivity::class.java)
     }
 }
